실수 가능하도록 추가

+++ Assets/Scripts/Character/MyCharacterComponent.cs
@@ -20,7 +20,7 @@ public class MyCharacterComponent : CharacterComponent, IBoardGameSubscriber
 		}
 	}
 
-	public IEnumerator OnRollDice(int diceCount, int nextBonusAddCount, int nextBonusMultiplyCount)
+	public IEnumerator OnRollDice(int diceCount, int nextBonusAddCount, float nextBonusMultiplyCount)
 	{
 		yield return null;
 	}
